chore: refactor deposit and withdraw button
