+++ src/components/_columns/StrategyMobileColumn.tsx
@@ -1,13 +1,10 @@
-import { Text, Tooltip } from "@chakra-ui/react"
+import { Text } from "@chakra-ui/react"
 import { PercentageText } from "components/PercentageText"
-import { BaseButton } from "components/_buttons/BaseButton"
-import { WithdrawButton } from "components/_buttons/WithdrawButton"
+import { DepositAndWithdrawButton } from "components/_buttons/DepositAndWithdrawButton"
 import { StrategySection } from "components/_tables/StrategySection"
 import { Timeline } from "data/context/homeContext"
 import { DepositModalType } from "data/hooks/useDepositModalStore"
 import { CellValue } from "react-table"
-import { analytics } from "utils/analytics"
-import { useAccount } from "wagmi"
 
 type StrategyMobileColumnProps = {
   timeline: Timeline
@@ -24,7 +21,6 @@ export const StrategyMobileColumn = ({
   timeline,
   onDepositModalOpen,
 }: StrategyMobileColumnProps) => {
-  const { isConnected } = useAccount()
   return [
     {
       Header: "Strategy",
@@ -55,47 +51,12 @@ export const StrategyMobileColumn = ({
     {
       Header: () => <Text>Deposit</Text>,
       id: "deposit",
-      Cell: ({ row }: any) => {
-        return (
-          <Tooltip
-            bg="surface.bg"
-            color="neutral.300"
-            label={
-              row.original.deprecated
-                ? "Strategy Deprecated"
-                : "Connect your wallet first"
-            }
-            shouldWrapChildren
-            display={
-              row.original.deprecated || !isConnected
-                ? "inline"
-                : "none"
-            }
-          >
-            <BaseButton
-              disabled={!isConnected}
-              variant="solid"
-              onClick={(e) => {
-                e.stopPropagation()
-                if (row.original.deprecated) {
-                  onDepositModalOpen({
-                    id: row.original.slug,
-                    type: "withdraw",
-                  })
-                  return
-                }
-                onDepositModalOpen({
-                  id: row.original.slug,
-                  type: "deposit",
-                })
-                analytics.track("home.deposit.modal-opened")
-              }}
-            >
-              {row.original.deprecated ? "Closed" : "Deposit"}
-            </BaseButton>
-          </Tooltip>
-        )
-      },
+      Cell: ({ row }: any) => (
+        <DepositAndWithdrawButton
+          row={row}
+          onDepositModalOpen={onDepositModalOpen}
+        />
+      ),
     },
   ]
 }
