+++ src/components/_columns/StrategyTabColumn.tsx
@@ -1,13 +1,9 @@
 import { Text, Tooltip, VStack } from "@chakra-ui/react"
 import { PercentageText } from "components/PercentageText"
-import { BaseButton } from "components/_buttons/BaseButton"
+import { DepositAndWithdrawButton } from "components/_buttons/DepositAndWithdrawButton"
 import { ApyRewardsSection } from "components/_tables/ApyRewardsSection"
 import { StrategySection } from "components/_tables/StrategySection"
 import { Timeline } from "data/context/homeContext"
-import { analytics } from "utils/analytics"
-import { useAccount } from "wagmi"
-import { formatDistanceStrict, isBefore } from "date-fns"
-import { zonedTimeToUtc } from "date-fns-tz"
 import { DepositModalType } from "data/hooks/useDepositModalStore"
 
 type StrategyTabColumnProps = {
@@ -25,8 +21,6 @@ export const StrategyTabColumn = ({
   timeline,
   onDepositModalOpen,
 }: StrategyTabColumnProps) => {
-  const { isConnected } = useAccount()
-
   return [
     {
       Header: "Strategy",
@@ -137,49 +131,12 @@ export const StrategyTabColumn = ({
     {
       Header: () => <Text>Deposit</Text>,
       id: "deposit",
-      Cell: ({ row }: any) => {
-        const date = new Date(row?.original?.launchDate as Date)
-        const dateTz = date && zonedTimeToUtc(date, "EST")
-        const isBeforeLaunch = isBefore(dateTz, new Date())
-        return (
-          <Tooltip
-            bg="surface.bg"
-            color="neutral.300"
-            label={
-              row.original.deprecated
-                ? "Strategy Deprecated"
-                : "Connect your wallet first"
-            }
-            shouldWrapChildren
-            display={
-              row.original.deprecated ||
-              !isConnected ||
-              !isBeforeLaunch
-                ? "inline"
-                : "none"
-            }
-          >
-            <BaseButton
-              disabled={
-                row.original.deprecated ||
-                !isConnected ||
-                !isBeforeLaunch
-              }
-              variant="solid"
-              onClick={(e) => {
-                e.stopPropagation()
-                onDepositModalOpen({
-                  id: row.original.slug,
-                  type: "deposit",
-                })
-                analytics.track("home.deposit.modal-opened")
-              }}
-            >
-              {row.original.deprecated ? "Closed" : "Deposit"}
-            </BaseButton>
-          </Tooltip>
-        )
-      },
+      Cell: ({ row }: any) => (
+        <DepositAndWithdrawButton
+          row={row}
+          onDepositModalOpen={onDepositModalOpen}
+        />
+      ),
     },
   ]
 }
